Backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==11.3.0
opencv-python==4.12.0.88
torch==2.7.1
torchvision==0.22.1
ultralytics==8.3.172
numpy==2.2.6
pandas==2.3.1
matplotlib==3.10.5
scipy==1.16.1
requests==2.32.4
PyYAML==6.0.2
tqdm==4.67.1
psutil==7.0.0 